More friendly scrolling
Earlier the user could not scroll before clicking the arrow or pressing the down key. But with this commit, he won't see the scroll bar as soon as the page will load to give a better look of the landing page and then either by pressing the down key or scrolling once, the scroll bar will appear.

Fixes #2